Add cascade mode interrupt support
Axi intc can be cascaded to psu_gic or ps7_gic. This patch
adds this support to generate proper interrupt ids and interrupt
parent info.
Ex: interrupt pin of a peripheral connected to axi_intc and axi
intc to gic. then interrupt parent of the peripheral is axi_intc
and parent to axi_intc is gic. and also axi_intc interrupt id
will be the id of pin connected to gic and the interrupt id
of the peripheral will be id connected to axi_intc.

axi_intc_0: interrupt-controller@80010000 {
	interrupt-controller ;
	interrupt-parent = <&gic>;
	interrupts = <0 89 4>;
};
axi_uart16550_0: serial@80130000 {
	interrupt-parent = <&axi_intc_0>;
	interrupts = <1 2>;
};

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Acked-by: Michal Simek <michals@xilinx.com>